refactor: extract helm utils to module
